% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattnlay.r
\docType{class}
\name{Layer-class}
\alias{Layer-class}
\alias{Layer}
\alias{r<-}
\alias{m<-}
\title{An S4 class to represent a layer around a nanoparticle}
\usage{
r(x) <- value

m(x) <- value
}
\description{
An S4 class to represent a layer around a nanoparticle
}
\section{Slots}{

\describe{
\item{\code{m}}{the complex refractive index}

\item{\code{r}}{the radius in nanometers}
}}

