% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattnlay.r
\docType{class}
\name{Scatterer-class}
\alias{Scatterer-class}
\alias{Scatterer}
\title{An S4 class to represent a scattering simulation}
\description{
An S4 class to represent a scattering simulation
}
\section{Slots}{

\describe{
\item{\code{na}}{the real refractive index of the ambient or surrounding media}

\item{\code{lambda}}{the wavelength of light in nanometers}

\item{\code{nt}}{the number of theta angles}

\item{\code{ti}}{the starting angle in degrees}

\item{\code{tf}}{the end angle in degrees}

\item{\code{layers}}{a list of layers of type Layer}
}}

